requirements.txt
numpy
selenium
onnx
onnxruntime
mongoengine
pydantic
pydantic-settings
tqdm
torchbnn